+++ gateway-service/src/main/java/com/ca/mfaas/gateway/error/custom/CustomErrorStatusHandlingBean.java
